refactor(label-sync): 提取 label 同步逻辑为可复用领域模块并自动触发
